An attempt to textGetRect faster (not much). Its much faster to use textPageGetRectsForRanges Tests for new apis

+++ pdfiumandroid/src/main/java/io/legere/pdfiumandroid/jni/NativeTextPage.kt
@@ -114,14 +114,14 @@ class NativeTextPage {
         private external fun nativeTextGetRect(
             textPagePtr: Long,
             rectIndex: Int,
-        ): DoubleArray
+        ): FloatArray
 
         @JvmStatic
         @FastNative
         private external fun nativeTextGetRects(
             textPagePtr: Long,
             wordRanges: IntArray,
-        ): DoubleArray?
+        ): FloatArray?
 
         @Suppress("LongParameterList")
         @JvmStatic
